+++ src/eca/config.clj
@@ -102,6 +102,7 @@
                             :key "${env:ANTHROPIC_API_KEY}"
                             :requiresAuth? true
                             :models {"claude-sonnet-4-6" {}
+                                     "claude-sonnet-5" {}
                                      "claude-opus-4-6" {}
                                      "claude-opus-4-7" {}
                                      "claude-opus-4-8" {}
@@ -206,7 +207,7 @@
               :shellCommand {:summaryMaxLength 35}
               :outputTruncation {:lines 2000 :sizeKb 50}}
    :variantsByModel {".*sonnet[-._]4[-._]6|opus[-._]4[-._][56]" {:variants anthropic-variants}
-                     ".*opus[-._]4[-._][78]|.*fable[-._]5|.*mythos[-._]5" {:variants anthropic-v2-variants}
+                     ".*opus[-._]4[-._][78]|.*sonnet[-._]5|.*fable[-._]5|.*mythos[-._]5" {:variants anthropic-v2-variants}
                      ".*gpt[-._]5(?:[-._](?:2|4|5)(?!\\d)|[-._]3[-._]codex)" {:variants openai-variants
                                                                               :excludeProviders ["github-copilot"]}
                      ".*deepseek[-._]v4[-._]pro" {:variants deepseek-variants